+++ src/babashka/impl/clojure/core.clj
@@ -31,6 +31,7 @@
 (def command-line-args (core-dynamic-var '*command-line-args*))
 (def warn-on-reflection (core-dynamic-var '*warn-on-reflection* false))
 (def compile-files (core-dynamic-var '*compile-files* false))
+(def unchecked-math (core-dynamic-var '*unchecked-math* false))
 (def math-context (core-dynamic-var '*math-context*))
 
 (defn read+string
@@ -169,6 +170,7 @@
    '*command-line-args* command-line-args
    '*warn-on-reflection* warn-on-reflection
    '*compile-files* compile-files
+   '*unchecked-math* unchecked-math
    '*math-context* math-context
    'with-precision (sci/copy-var with-precision clojure-core-ns)
    '-with-precision (sci/copy-var -with-precision clojure-core-ns)
@@ -193,5 +195,7 @@
    'abs (sci/copy-var abs clojure-core-ns)
    'StackTraceElement->vec (sci/copy-var StackTraceElement->vec clojure-core-ns)
    'memfn (sci/copy-var memfn clojure-core-ns)
-   'into-array (sci/copy-var into-array clojure-core-ns)}
+   'into-array (sci/copy-var into-array clojure-core-ns)
+   'print-method (sci/copy-var print-method clojure-core-ns)
+   'print-dup (sci/copy-var print-dup clojure-core-ns)}
   )