+++ src/babashka/impl/process.clj
@@ -1,6 +1,7 @@
 (ns babashka.impl.process
   {:no-doc true}
   (:require [babashka.process :as process]
+            [babashka.process.pprint]
             [sci.core :as sci :refer [copy-var]]))
 
 (def tns (sci/create-ns 'babashka.process nil))
@@ -26,4 +27,5 @@
    'tokenize    (copy-var process/tokenize tns)
    '*defaults*  defaults
    'destroy     (copy-var process/destroy tns)
-   'destroy-tree (copy-var process/destroy-tree tns)})
+   'destroy-tree (copy-var process/destroy-tree tns)
+   'exec (copy-var process/exec tns)})