+++ test-resources/babashka/uberscript/src/my/main_pod.clj
@@ -1,0 +1,6 @@
+(ns my.main-pod
+  (:require [my.other-ns-with-pod]
+            [pod.babashka.go-sqlite3 :as sqlite]))
+
+(defn -main [& _args]
+  (sqlite/query ":memory:" ["SELECT 1 + 2 as sum"]))